Begin implementing jwt login authentication
Successfully verifying a sent password matches the stored db hash. I'm
assuming hashing will be done on the client side (so plain text
passwords are not sent accross the web to api) but I don't think this is
implimented yet. Currently I am returning the accessToken as a cookie,
we'll have to make sure that is used properly on the react app.